Refactor: modularise Flask app with Blueprints and factory pattern
