+++ init_db.py
@@ -1,6 +1,10 @@
-from run import app
-from models import db
+# init_db.py
+
+from app import create_app
+from app.models import db
+
+app = create_app()
 
 with app.app_context():
     db.create_all()
-    print("✅ Database initialized successfully!")
+    print("✅ All tables created in the database.")
