+++ run.py
@@ -1,43 +1,21 @@
-#!/usr/bin/env python3
-"""
-Main application entry point for the Spotify Mood Analysis application.
-"""
-from dotenv import load_dotenv
-load_dotenv()
+# run.py
 
 import os
+import sys
 
-try:
-    from .config import config
-    from app import create_app
-except ImportError:
-    # Allow running as a standalone script during development
-    from config import config
-    # from __init__ import create_app
+sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 
+from dotenv import load_dotenv
 
-# Create the app globally so it can be imported
-config_name = os.environ.get('FLASK_CONFIG', 'development')
+# Load environment variables from .env file
+load_dotenv()
+
+from app import create_app
+
+# Load config name from env, default to 'development'
+config_name = os.getenv('FLASK_CONFIG', 'development')
 app = create_app(config_name)
 
-def main():
-    """
-    Run the application using the configuration from environment variables.
-    """
-    try:
-        # # Get configuration from environment variable or use default
-        # config_name = os.environ.get('FLASK_CONFIG', 'development')
-
-        # # Create app with the proper configuration
-        # app = create_app(config_name)
-
-        # Run with debugging enabled or disabled based on configuration
-        app.run(debug=app.config['DEBUG'])
-    except ImportError:
-        print("\nMissing required packages! Please run:")
-        print("   pip install -r requirements.txt\n")
-        exit(1)
-
-
 if __name__ == '__main__':
-    main()
+    port = int(os.environ.get('PORT', 5000))
+    app.run(host='0.0.0.0', port=port, debug=app.config['DEBUG'])
