+++ templates/login.html
@@ -39,7 +39,7 @@
         <h2>Login</h2>
 
         <!-- Flask-WTF form begins here with CSRF token automatically added -->
-        <form method="POST" action="{{ url_for('login') }}">
+        <form method="POST" action="{{ url_for('user.login') }}">
             {{ form.hidden_tag() }}
 
             <!-- Email input field: floating label effect handled via CSS transitions -->
@@ -63,14 +63,14 @@
         <!-- Links for password recovery and new user sign-up -->
         <div class="group">
             <a href="#">Forgot Password</a>
-            <a href="{{ url_for('signup') }}">Sign up</a>
+            <a href="{{ url_for('user.signup') }}">Sign up</a>
         </div>
 
         <!-- Visual divider for alternate login methods -->
         <div class="or-divider"><span>OR</span></div>
 
         <!-- Spotify OAuth login button with redirect using Flask route -->
-        <button type="button" class="inputBox oauth-btn spotify-btn" onclick="location.href='{{ url_for('oauth_spotify') }}'">
+        <button type="button" class="inputBox oauth-btn spotify-btn" onclick="location.href='{{ url_for('spotify.oauth_spotify') }}'">
             <img src="https://storage.googleapis.com/pr-newsroom-wp/1/2023/05/Spotify_Primary_Logo_RGB_Green.png"
                 alt="Spotify logo" class="btn-icon">
             Connect with Spotify
