+++ templates/signup_cred.html
@@ -54,7 +54,7 @@
         <h2>Sign Up</h2>
 
         <!-- Credentials Form -->
-        <form method="POST" action="{{ url_for('signup_login_credentials') }}">
+        <form method="POST" action="{{ url_for('user.signup_login_credentials') }}">
             {{ form.hidden_tag() }}
 
             <!-- Email Input Field -->
@@ -83,14 +83,14 @@
 
         <!-- Links for password recovery and new user sign-up -->
         <div class="group">
-            <a href="{{ url_for('login') }}">Already have an account? Login</a>
+            <a href="{{ url_for('user.login') }}">Already have an account? Login</a>
         </div>
         
         <!-- Visual divider for alternate login methods -->
         <div class="or-divider"><span>OR</span></div>
         
         <!-- Spotify OAuth login button with redirect using Flask route -->
-        <button type="button" class="inputBox oauth-btn spotify-btn" onclick="location.href='{{ url_for('oauth_spotify') }}'">
+        <button type="button" class="inputBox oauth-btn spotify-btn" onclick="location.href='{{ url_for('spotify.oauth_spotify') }}'">
             <img src="https://storage.googleapis.com/pr-newsroom-wp/1/2023/05/Spotify_Primary_Logo_RGB_Green.png"
                         alt="Spotify logo" class="btn-icon">
                     Connect with Spotify
